Introduce volume correction
IAudioEndpointVolume is now used to read and set mute state and levels.
The sending of volume up/down is kept as fall-back procedure.
Should for some reason the correction not work,so is it frequently
amendable by reloading addons (Control+NVDA+F3).
The GUI does currently not reflect the enhanced possibilities.