Add contextual help
It is now possible to press F1 in any control of the settings dialog
to show a browseable message with an automatic  help extract for that control.
In order to do this correctly, the focus behaviour of the dialog has been changed.
Formerly, switching forth and back to the dialog focused always on the first element.
It now returns to the right child control.
(issue/PR for NVDA should be created for this behaviour)